+++ src/sentry/static/sentry/app/redux/actions/workBatchDetails_old.js
